feat(java): make work in docker

+++ services/Cart_java/src/main/java/dev/ork/Cart/Cart/CartController.java
@@ -3,19 +3,19 @@ package dev.ork.Cart.Cart;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/cart")
-
 public class CartController {
+
   @Autowired
   private CartRepository cartRepository;
 
@@ -24,32 +24,35 @@ public class CartController {
     return cartRepository.findAll();
   }
 
+  @SuppressWarnings("null")
   @GetMapping("/{id}")
   public Cart getCartById(@PathVariable Long id) {
-    return cartRepository.findById(id).get();
+    return cartRepository.findById(id).orElse(null);
   }
 
+  @SuppressWarnings("null")
   @PostMapping
-  public Cart createCart(@RequestBody Cart productID) {
-    return cartRepository.save(productID);
+  public Cart createCart(@RequestBody Cart cart) {
+    return cartRepository.save(cart);
   }
 
+  @SuppressWarnings("null")
   @PutMapping("/{id}")
-  public Cart updateCart(@PathVariable Long id, @RequestBody Cart productID) {
-    Cart existingCart = cartRepository.findById(id).get();
-    existingCart.setProductId(productID.getProductId());
-    return cartRepository.save(existingCart);
+  public Cart updateCart(@PathVariable Long id, @RequestBody Cart updatedCart) {
+    return cartRepository.findById(id).map(cart -> {
+      cart.setProductId(updatedCart.getProductId());
+      return cartRepository.save(cart);
+    }).orElse(null);
   }
 
+  @SuppressWarnings("null")
   @DeleteMapping("/{id}")
   public String deleteCart(@PathVariable Long id) {
-    try {
-      cartRepository.findById(id).get();
+    if (cartRepository.existsById(id)) {
       cartRepository.deleteById(id);
       return "Cart deleted successfully";
-    } catch (Exception e) {
+    } else {
       return "Cart not found";
     }
   }
-
 }
